Fixed a bug with Round Robin and if process was added before context switch was complete
There was a bug where the Round Robin would never add in a process if
there was 1 process left. Also there was a bug if the process was added
before the context switch of a processor was complete then the process
would disregard the time limit it had on itself or processor.