Implemented Visual Display and made alot of changes
A few bug fixes with removing and adding processes. The main pane was
implemented, which is a visual display of where each process is located.
The processes are represented as circles and move based one where they
are in the simulation. Also redid how the processes are randomly
generated with their arrival time. Instead of generating the arrival
time in the strategy, it is generated like the rest of the data values
and the processes are created right after. Fixed a few validation bugs
like if the user tried to add in negative values.