Implemented IO from a file
The functionallity for input from a file has always been in the project,
although it was never accessed by the system. The system can now access
data from a text file and load the saved processes into the process
editor. The user can also save the data to the same file.