Added Text To Processes and Fixed bug with Random Generated
Added text in the middle of the circles of each process. If the text is
longer then the circle it will bleed over. This is a minor bug that
could be fixed. A major bug was fixed when randomly generated processes
were created. If the CPU time is greater then the process time. It would
run the CPU time reqardless. Small if statement change fixed the problem